reorg server
